UpdatedReadme.adoc
I tried following the steps given but could not successfully execute the project using IntelliJ Community Edition 2016. So, when I started experimenting, found that for this example to work as expected, I had to choose "Use gradle wrapper task configuration" without which the gradle.wrapper.properties file and gradle-wrapper.jar are not created. As I am new to Spring and IntelliJ, kindly let me know if I am not making much sense. Thanks.